Seals sales was finished

+++ app/src/main/java/com/qnecesitas/elretenretenes/data/Sales.kt
@@ -6,12 +6,12 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class Sales(
-    @PrimaryKey(autoGenerate = false) var code:String ,
-    @ColumnInfo(name = "codeProduct") var codeProduct:String ,
-    @ColumnInfo(name = "totalPrice") var totalPrice:Int ,
-    @ColumnInfo(name = "totalInv") var totalInv:Int ,
-    @ColumnInfo(name = "amount") var amount:Int ,
-    @ColumnInfo(name = "day") var day:Int ,
-    @ColumnInfo(name = "month") var month:Int ,
-    @ColumnInfo(name = "year") var year:Int
-    )
+    @PrimaryKey(autoGenerate = false) var code: String ,
+    @ColumnInfo(name = "codeProduct") var codeProduct: String ,
+    @ColumnInfo(name = "totalPrice") var totalPrice: Double ,
+    @ColumnInfo(name = "totalInv") var totalInv: Double ,
+    @ColumnInfo(name = "amount") var amount: Int ,
+    @ColumnInfo(name = "day") var day: Int ,
+    @ColumnInfo(name = "month") var month: Int ,
+    @ColumnInfo(name = "year") var year: Int
+)
